// web/src/contracts/event.schema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Event {
  t_ms: number;
  type:
    | "gaze"
    | "fixation"
    | "cursor_dwell"
    | "hover"
    | "pickup"
    | "add_to_cart"
    | "remove"
    | "station_enter"
    | "station_exit"
    | "checkout";
  station_id: string | null;
  payload: {
    [k: string]: unknown;
  };
}